Rename map explorer file
